lib/9pfs: Provide `TIOCGWINSZ`-aware `ioctl` stub
`python3` uses `isatty()`, which does an `ioctl()` with
`TIOCGWINSZ` as an argument for checking whether it should invoke
the interactive interpretor or just start interpreting a given
file.

Since `9pfs` is not a pseudo filesystem, it should be able to respond
to such `ioctl` accordingly for scripts defined in its filesystem
with a `ENOTTY`.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>