lib/vfscore: Fix wrong application of umask
Previously vfscore would apply the umask to the mode bits in the
emulated libc functions `open` and `openat`, instead of within the
open/openat syscall implementation where it would be correct to do so.
This lead to the umask being ignored when a real libc was in use.
This change fixes this bug, restoring expected behavior with or without
a libc.

Signed-off-by: Andrei Tatar <andrei@unikraft.io>