lib/ukconsole: Add library for serial console
This commit migrates the console API into a new libukconsole.
Each console provides an ops structure that implements the console API.
The APIs include coutk, coutd, cink, and init.

A unified uk_console_init function takes bootinfo as a parameter to
initialize the available console in both x86 and ARM platforms. The
implementation of uk_console_init is platform-specific.

The header for ANSI characters is moved into a separate one in
include/uk.

Signed-off-by: Xingjian Zhang <zhxj9823@qq.com>