.github/workflows: Introduce GitHub Actions Workflow integrations
This commit introduces GitHub-centric CI/CD definition that is intended
to be executed on incoming PRs such that the incoming PR can be
determined as working.  This supplementary file is based on-top of the
KraftKit GitHub Action Workflow Step[0] which is a native-to-Unikraft
integration.  This integration allows for seamless building, packaging,
distributing and executing (via TCG-based emulation) of Unikraft
unikernels in a clear and declarative manner.

In this first integration pipeline, incoming PRs will be built against
the following variant of the helloworld application:

- qemu/x86_64
- qemu/arm64
- xen/x86_64
- fc/x86_64
- linuxu/x86_64
- linuxu/arm64

Additionally, a more comprensive test will be executed with libc-test[1]
via uktest in emulation mode for qemu on x86_64 only in order to display
maximum coverage.

[0]: https://github.com/unikraft/kraftkit/blob/staging/action.yml
[1]: https://github.com/unikraft/lib-libc-test

Signed-off-by: Alexander Jung <alex@unikraft.io>
Reviewed-by: Cezar Craciunoiu <cezar.craciunoiu@unikraft.io>
Approved-by: Razvan Deaconescu <razvand@unikraft.io>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #1012