lib/ukdebug: Add colorless `LVLC_THREAD` macro
Choosing to print thread identifier fails at build time when the
`CONFIG_LIBUKDEBUG_ANSI_COLOR` option is not selected.
This is due to the fact that `LVLC_THREAD` is defined only with
colored output.

As a result, we define this macro with an empty string even when
`CONFIG_LIBUKDEBUG_ANSI_COLOR` is not set.

Signed-off-by: Maria Sfiraiala <maria.sfiraiala@gmail.com>