lib/nolibc: Provide symbols for `malloc`, `free`, `calloc`, `realloc`
`posix_memalign` and `memalign`

To ensure that symbols such as malloc, calloc, and free are accessible
during link time for internal gcc objects, such as those related to gcov,
these functions must be designated as public.

As a side-effect of this change, unikraft loses the fine-grained stats
introduced in #229 when using the before-mentioned calls,
and instead each operation via the libc interface will be accounted
to `nolibc` instead to the individual libraries that made the calls.

Signed-off-by: Eduard-Florin Mihailescu <mihailescu.eduard@gmail.com>
iReviewed-by: Andra Paraschiv <andra@unikraft.org>
Reviewed-by: Radu Nichita <radunichita99@gmail.com>
Approved-by: Simon Kuenzer <simon@unikraft.io>
GitHub-Closes: #1042