plat/common: Add BIOS system memory in legacy high memory region
Previously, `ukplat_memregion_list_insert_legacy_hi_mem` did not
include the `0xf0000 -> 0x100000` memory regions that would usually
contain the legacy BIOS system memory below the first 1MiB. This
would lead to x86 platforms that do not yet have any form of
ACPI (RSDP from BDA or UEFI System Tables) such as Firecracker
to crash when looking for the `RSDP` in that specific region.

Therefore, make sure that this region is also included so that it
can be mapped accordingly by the paging initialization phase.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>