plat/common: Align TLS sections to PAGE_SIZE
`mkbootinfo.py` takes each ELF segment and aligns it to PAGE_SIZE to
convert it to a `struct ukplat_memregion_desc`. However, this ends up
generating overlapping memregion descriptors in the case of the ELF
segment corresponding to the TLS sections.

To solve this, simply ensure that the TLS sections are already aligned
by PAGE_SIZE.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>