plat/kvm: Use `AllocatePages()` to allocate memory regions in `UEFI`
According to the `UEFI` specifications, `AllocatePool()` is the safest
allocation method, however there are two pitfalls:
- we don't know the BIOS's memory allocator or the system's memory map,
so the resulted allocation start address may be out of the range that
would be compatible with a Unikraft configuration that does not have
`CONFIG_HAVE_PAGING`
- it does not guarantee page-aligned allocations and, even though we
may be able to align the length or the end address, the allocation's
start address may therefore not be page-aligned

Thus, replace such calls to `AllocatePool()` with calls to
`AllocatePages()` to ensure page-aligned start addresses. Furthermore,
preserve backwards compatibility with Unikraft configurations that do
not have `CONFIG_HAVE_PAGING` by limiting the allocation's start address
to be below the highest address of the static boot page tables through
the help of `bpt_unmap_mrd`.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>