plat/common: Preserve original size of allocated memory regions
Previously, `ukplat_memregion_alloc()` would allow one to dynamically
allocate based on available free memory regions a certain memory region
of a specific type and with specific flags and whose size will be the
page-aligned size of the desired size.

A user may want however to fetch some memory regions and know the
actual, intended size. To achieve this, make it so that the allocated
memory region has the original size on the surface, but the actual
allocation is done by splitting a page-aligned chunk to avoid
misalignments.

This leads to the same amount of memory fragmentation as before,
but with the improvement that the user of the API will have access to
the actual size of the data the memory region was allocated for.

Luckily, the paging initialization code is written in such a way that
it aligns every memory region before mapping/unmapping.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>