plat/kvm: Change the bootflow to adapt to Arm CCA
This commit adds uk_rsi_init function to initialize the RSI interface.
The memory regions for the realm is also initialized at the beginning of
_ukplat_entry. When libukrsi is enabled, console initialization is
deferred after the initialization of paging, which additionally setup
the device region.

It also changes the size of the boot stack to 16KB to accommodate
larger objects used in the boot flow.

Signed-off-by: Xingjian Zhang <zhxj9823@qq.com>