plat/common: Setup realm memory region for CCA
The unprotected alias of the device region is excluded from unmapping.
The handling of unmapping regions is changed as not only the first two
regions in the list would be UKPLAT_MEMRF_UNMAP.

The device region are setup at the end of ukplat_mem_init, where the
entire device region are marked as unprotected. The device region is
determined by new macros defined in kvm-arm64/image.h.

Signed-off-by: Xingjian Zhang <zhxj9823@qq.com>