drivers/ukbus/pci: Fix result check of drv->add_dev() operation
Fix a bug in pci_ecam where a check against the result of drv->add_dev()
would happen outside the scope of the conditional call to that function.

This would trigger failure of arch_pci_driver_add_device() depending on
the state of the stack.

This bug was pinpointed by a long ignored GCC warning:

  warning: ‘ret’ may be used uninitialized in this function [-Wmaybe-uninitialized]

Checkpatch-Ignore: COMMIT_LOG_LONG_LINE
Signed-off-by: Michalis Pappas <michalis@unikraft.io>