support/scripts: `checkpatch` `WARN` on commit subject > 70 chars
Although `checkpatch.pl` checks for commit message lines to not be
longer than `75` characters, in the case of `Github`'s UI, a commit
subject longer than `70` character results in its line overflowing
into the next.
E.g.
```
[Selector]/[Component]: Lorem ipsum dolor sit amet, consectetuer adipiscin
```
becomes
```
[Selector]/[Component]: Lorem ipsum dolor sit amet, consectetuer adip...
...iscin
```

Since this is not aesthetically pleasing, enforce a character limit for
commit message subjects to not be longer than `70` characters.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>