build: Always save architecture with savedefconfig
With `savedefconfig`, KConfig creates a file that contains only options
that are different from the default selection. Because the build system
automatically detects the host architecture and sets the default
architecture to this detected architecture, the architecture is never
part of the defconfig file if the architecture in the configuration
matches the host architecture. This becomes problematic as soon as such
a file is loaded on a build environment that has a different host
architecture: The configuration can fail because of architectural
configuration differences.

This commit ensures that the configured architecture is always stored
in the defconfig file. The `savedefconfig` target appends the configured
architecture to the file if the architectures between host and config
match.

Signed-off-by: Simon Kuenzer <simon@unikraft.io>