build: Rename `_import_lib` to `import_lib`
This commit renames the Makefile function `_import_lib` to `import_lib`.
The original intent was for this function to be purely internal and only
used in the core `Makefile`. It turns out that it needs to be used in many
other places, like `drivers/`, so this commit removes the inappropriate
`_` prefix.

Signed-off-by: Simon Kuenzer <simon@unikraft.io>