build: Rename `_import_linker` to `import_linker`
This commit renames the Makefile function `_import_linker` to `import_linker`.
The original intent was for this function to be purely internal and only
used in the core `Makefile`. It turns out that it might be used in other
places in the future, like potentilly external architectures, so this
commit removes the inappropriate `_` prefix.

Signed-off-by: Simon Kuenzer <simon@unikraft.io>