build: Forward YACC/LEX to kconfig make invocation
KConfig relies on GNU bison and does not work with the classical yacc.
While most Linux distributions prefer bison and add a symlink for yacc,
there are distributions that do not this (Fedora). They seem to actually
use the original Berkeley yacc to provide `yacc`.

Instead of using the default `yacc` for `YACC` in make explicitly forward
our preference for `bison` to the kconfig Makefile.

Signed-off-by: Marco Schlumpp <marco@unikraft.io>
Reviewed-by: Stefan Jumarea <stefanjumarea02@gmail.com>
Reviewed-by: Maria Sfiraiala <maria.sfiraiala@gmail.com>
Approved-by: Razvan Deaconescu <razvand@unikraft.io>
GitHub-Closes: #1092