lib/uklibid: Embedded library information
This commit introduces the `.uk_libinfo` section which is used to store
additional library information in library object files. We define the data
structure of the section and declare it as layout version 1. For each
library the build system will generate a single header and each header can
contain an individual number of records to store values.
Layout version 1 also supports storing of information related to the image
as a whole and not to a specific library. A single header is created for
global information, which differs from the library headers in that it does
not contain a record for a library name (LIBNAME).

When a final Unikraft image is linked, the individual sections of each
library are merged into one section.

By default, a minimal set of library and global information is enabled. The
motivation behind this is that library objects are instrumented with
uk_libinfo from now on. In a future commit, the build system could provide
an option to remove this section from (non-debug) images if it is not used
by Unikraft code itself.

Signed-off-by: Simon Kuenzer <simon@unikraft.io>