build: Add target to print unikernel LoC stats
This change adds the `print-loc` Makefile target which (fairly hackishly)
computes and prints statistics about the number of lines of code that
make it into a built unikernel image. It is not intended as an end-user
metric and thus does not aim for any particular level of accuracy or
edge-case handling, rather it is a fun little dev tool that allows
rough comparisons of the level of "bloat" between unikernels.

Signed-off-by: Andrei Tatar <andrei@unikraft.io>
Reviewed-by: Stefan Jumarea <stefanjumarea02@gmail.com>
Approved-by: Razvan Deaconescu <razvand@unikraft.io>
GitHub-Closes: #1108