build: Introduce `E=` parameter for excluding libraries
This commit introduces the build system parameter `E=` which expects a
colon separated list of paths that are matched against any library path
that the build system is including. This mechanism is intended to be used
for replacing internal libraries.
For example, to replace `uksched`, the replacing library (that has to match
configuration name(s) and APIs) needs to be included with `L=` while at the
same time, the Unikraft-internal library has to be excluded with `E=`:
```
 make L=/path/to/external-uksched E=/path/to/unikraft/lib/uksched
```

Signed-off-by: Simon Kuenzer <simon@unikraft.io>