plat/common/memory: Include `EBDA` in default legacy high memory
The RSDP of ACPI or the MP Floating Pointer Structure of the MP
specification may be found in the first 1KB of EBDA. This is
valuable for very old systems complying to PC-AT multiprocessor
systems. Thus, make sure we include this valuable memory portion
in our hardcoded, x86 specific, legacy high memory regions.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>