plat/common/memory: Introduce `UKPLAT_MEMRT_LBIOS`
On x86, there is a special, predetermined place for legacy BIOS related
memory regions such as the EBDA, the XROMs read by the BIOS from add-on
cards or the BIOS System ROM itself.

Since these are frequently used when parsing important boot time
information related to specifications such as ACPI or the MP
specification, dedicate a special memory region type to these important
regions.

Furthermore, add a case for such region in `get_mrd_prio` and give it
the priority of reserved memory regions.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>