build: Fix rule that copies `.config` to `build/config`
The rule that copies the current `.config` to the build directory was
intended, among other things, to detect configuration changes. However, the
corresponding rule always copies the file, causing to update the timestamp
on every make call. This commit fixes that rule by only copying the file
when there have been configuration updates.

Signed-off-by: Simon Kuenzer <simon@unikraft.io>