lib/uklibid: Store compile options to `.uk_libinfo`
This commit introduces the storage of an initial set of compile options in
the metadata of each library. The following options are covered by this
commit:
- PIE: Library was compiled position-independent
- DCE: Library was compiled with support for dead code elimination
- LTO: Library was compiled with support for link time optimizations
The options are always saved, because some of the activation states cannot
be mixed, such as the PIE option (e.g., an Unikraft build must be pure PIE
enabled or pure PIE disabled) It is also intended that the encoded flags
are extended in the future.

Signed-off-by: Simon Kuenzer <simon@unikraft.io>
Reviewed-by: Michalis Pappas <michalis@unikraft.io>
Reviewed-by: Robert Kuban <robert.kuban@opensynergy.com>
Approved-by: Razvan Deaconescu <razvand@unikraft.io>
GitHub-Closes: #1117