lib/uklibid: Option to store "Compile date" to `.uklibid_info`
Adds the option to store the compile time and date for a library and
unikernel. When enabled, `libinfo.S` is recompiled on every rebuild. This
is achieved by handing over the current time and date as compile flag.
For this purpose, the build system provides `$(HOSTUTC)`, the current
time and date in UTC format.

Signed-off-by: Simon Kuenzer <simon@unikraft.io>