arch/*: Update and align paddr/vaddr validation on x86 and arm64
Changed the signature of paddr and vaddr range check to takes the
length of the range instead of the end address, which is more
clear and less error-prone.
The end address is subtracted by 1 to account for the maximum
address in the range.

Also changed the use sites of address range validation to use the
updated signature.

Signed-off-by: Kha Dinh <khadinh@g.skku.edu>
Reviewed-by: Sergiu Moga <sergiu@unikraft.io>
Reviewed-by: Michalis Pappas <michalis@unikraft.io>
Reviewed-by: Razvan Virtan <virtanrazvan@gmail.com>
Approved-by: Razvan Deaconescu <razvand@unikraft.io>
GitHub-Closes: #1091