lib/uklibid: Clang-compatible compile option flags
Clang's preprocessor seems to behave differently when computing an integer
based on undefined macros. For example, the following expression:
```
  (!(!__LIBUKLIBID_COMPILEOPT_PIE__) * (0x01 << 0))
```
resolves to 0x0 on GCC's GAS if __LIBUKLIBID_COMPILEOPT_PIE__ is not
defined. Clang is throwing an compile error because of the missing
definition. This commit adopts `libinfo.S` to define an intermediate
value macro that is set to `0` when the requested feature is not enabled.

Signed-off-by: Simon Kuenzer <simon@unikraft.io>