plat/kvm/x86: Add formatted printing utility using during EFI boot
The current crash printing is somewhat limited as it cannot print
variables.
This commit refactor EFI debug printing. It adds a printf-like
utility for easier time debugging, rename some functions to be
more consistent with the rest of the code base.
Also, debug printing in EFI now depends on `libukdebug`, instead of
a dedicated option.

Signed-off-by: Kha Dinh <khadinh@g.skku.edu>