lib/ukboot: Handle shutdown requests with "init" thread
This commit introduces a "main" thread for calling the application's
`main()` function. This is done so that "init" is purely responsible
for system initialization and shutdown. After creating the "main" thread,
"init" will block on a barrier which can only be unblocked by a
call to `uk_boot_shutdown_req()`. "init" will then perform a system
shutdown. With this model, any thread, including "main", can request
a shutdown.

Signed-off-by: Simon Kuenzer <simon@unikraft.io>
Reviewed-by: Sergiu Moga <sergiu@unikraft.io>
Approved-by: Razvan Deaconescu <razvand@unikraft.io>
GitHub-Closes: #1148