build: Forcefully re-create link to main Makefile
This commit makes sure that the symbolic link within the build directory
always points to the unikraft Makefile version that is used for a
build. This avoids issues when a project is build with multiple Unikraft
versions but using the same build path. The reason is that the build system
is invoking a sub-make call within the build directory using this symbolic
link.

Signed-off-by: Simon Kuenzer <simon@unikraft.io>