lib/vfscore: Add `mkmp` to `ukopts` as a `vfs.fstab` option
In case someone may want to mount a filesystem to a directory
that does not currently exist in the rootfs, allow them to do so
without having to rebuild the rootfs by exposing a `ukopts`
option called `mkmp`. This option will let `fstab` know that
it must ensure that the mount path exists before attempting to
mount.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>