lib/vfscore: Introduce `ukopts`, a new `vfs.fstab` option
Allow the `vfs.fstab` option arguments to receive an additional
option that is Unikraft specific and that enables one to change
a volume's mount behavior depending on the passed option.

Thus, now the `vfs.fstab` syntax is as follows:
```
 	vfs.fstab=[
 		"<src_dev>:<mntpoint>:<fsdriver>[:<flags>:<opts>:<ukopts>]"
 		"<src_dev>:<mntpoint>:<fsdriver>[:<flags>:<opts>:<ukopts>]"
 		...
 	]
```

The `ukopts` options will have the same separator as `fstab`, that
being `:`.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>
Reviewed-by: Simon Kuenzer <simon@unikraft.io>
Approved-by: Simon Kuenzer <simon@unikraft.io>
GitHub-Closes: #1145