plat/common: Add `auxsp` field to `struct lcpu`
`struct uk_thread` contains a field called `auxsp` which is meant to
represent an auxiliary stack that may be used when switching stacks
during syscall entries or when simply wanting to have a scratch space
to use during a very fragile state of the system, such as when handling
an exception.

Give access to this field through `struct lcpu` as well. This field
shall represent the auxiliary stack pointer of the thread currently
executing on this LCPU.

If uksched/ukthread is disabled, this can come in handy as simply a
pointer to a scratch space that we can contaminate with whatever we
may want during a more fragile execution context.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>
Reviewed-by: Michalis Pappas <michalis@unikraft.io>
Approved-by: Razvan Deaconescu <razvand@unikraft.io>
GitHub-Closes: #1174