plat/common/x86: Switch to auxiliary stack on syscall entry
We would usually use the application's stack to execute the issued
system call. However this present a couple of problems such as
if the application's stacks are too small (like it is the case
for Go's goroutines) this will end up generating an unhandled
pagefault or, even worse, corrupting other memory areas beyond
the respective stack. Another problem would be that we will
end up overwriting stuff on the Red Zone for those applications
that employ it.

To fix this, switch to the per-thread auxiliary stack on system call
entry and use that instead. The switch shall be done with the help of
the `swapgs` instruction to help us swap the `gs_base` registers, as
our own `KERNEL_GS_BASE` contains a pointer to the current LCPU's
`struct lcpu` that has an up to date pointer to the current thread's
auxiliary stack.

Co-authored-by: Marco Schlumpp <marco@unikraft.io>
Signed-off-by: Marco Schlumpp <marco@unikraft.io>
Signed-off-by: Sergiu Moga <sergiu@unikraft.io>