plat/kvm/arm: Switch to auxiliary stack on syscall entry
We would usually use the trap stack to execute the issued
system call. However this may be troublesome if the syscall will
generate page faults for example, which would end up reusing the
same stack.

To fix this, switch to the per-thread auxiliary stack on system call
entry and use that instead. The switch shall be done with the help of
the `tpidr_el1` register to help us, as it will be guaranteed to contain
`struct lcpu` that has an up to date pointer to the current thread's
auxiliary stack.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>