plat/common/x86: Save `ss`/`rsp`/`eflags`/`cs`/`rip` on syscall frame
Previously, we would save zeroes on the syscall entry frame instead
of using the actual values for the registers that would normally be
popped off the stack since they would not be used anyway by a syscall.

For completeness's sake, do still save them, to offer those that may
view the structure a clear view of what the syscall will return to.

Furthermore, we no longer need to clear the stack before syscall exit
(the previous `addq $(6 * 8), %rsp`) because the auxiliary stack is
used instead, so do not do this operation anymore.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>
Reviewed-by: Michalis Pappas <michalis@unikraft.io>
Approved-by: Razvan Deaconescu <razvand@unikraft.io>
GitHub-Closes: #1174