lib/posix-process: Copy full parent registers context on `clone`
Register `clone` through `UK_LLSYSCALL_R_U_DEFINE` to have full
access to the `struct uk_syscall_ctx` context of the caller/parent.
This way we will be able to give the child the same exact register
context as the parent, except, of course, for the obvious registers
(e.g. `%rax` equal to `0` on `x86_64` and `x0` equal to `0`
on `ARM64`).

Some runtimes, such as the Go one, may expect some registers
to be exact the same as those of the parent.

Registers that may not be preserved may be the scratch ones
(e.g. `%r11` on x86).

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>