build: Introduce `UK_IMAGE_NAME_OVERWRITE` make var
This commit introduces a new make variable,
`UK_IMAGE_NAME_OVERWRITE`, which is used to customize the
output binary name of the kernel image.  This is a special,
and experimental, variable which can be helpful in certain
use cases, for example where targets with the same architecture
and platform are present.  This variable should be used in
conjunction with a different build directory, `O=`, as the
resulting binary will always have the name from this option.

Signed-off-by: Alexander Jung <alex@unikraft.io>
Reviewed-by: Simon Kuenzer <simon@unikraft.io>
Approved-by: Simon Kuenzer <simon@unikraft.io>
GitHub-Closes: #1169