arch/arm64: Implement stack switching on IRQ/trap
Define a per-cpu buffer whose size is twice that of the configured
stacks and whose definition may be represented through the following
diagram:
```
      STACK_SIZE               STACK_SIZE
 <---------------------><--------------------->
 |============================================|
 |                     |                      |
 |       trap stack    |        IRQ stack     |
 |                     |                      |
 |=============================================
                       ^
                     SP_EL0
```

The middle address of this buffer shall be assigned to `SP_EL0`,
a great candidate for a register free of use as a `EL1` only Unikernel.

Now, depending on whether an exception is an IRQ or a trap, the early
assembly entry will switch to either the IRQ stack or the trap stack,
by simply making use of the `SP_EL0` system register.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>