include/essentials: Provide `__naked` macro
Add a macro for `__attribute__((naked))`. This hints to the compiler
to generate code without prolog and epilog code. This comes in handy
when writing inline assembly functions only. Although caution may
be required as the ABI must still be respected and the caller
registers that are to be preserved by the callee must still be
preserved.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>