lib/ukcpio: Overwrite existing files
This commit changes the behavior of `ukcpio_extract()` so that any file
that may already exist is overwritten during extraction. With this change,
we can implement a sort of overlaying a file system (e.g. an embedded
initrd) with the contents of an external initrd by overwriting the original
contents with new files.
Because CPIOs are extracted to an empty ramfs mount point in most cases,
this change has no effect on normal initrd usage.

Checkpatch-Ignore: AVOID_EXTERNS
Checkpatch-Ignore: FUNCTION_ARGUMENTS
Signed-off-by: Simon Kuenzer <simon@unikraft.io>
Approved-by: Razvan Deaconescu <razvand@unikraft.io>
Reviewed-by: Stefan Jumarea <stefanjumarea02@gmail.com>
Reviewed-by: Sergiu Moga <sergiu@unikraft.io>
GitHub-Closes: #1201